[FEAT] spring data jpa, controller

+++ src/main/java/com/excilys/cdb/dto/ComputerDTO.java
@@ -56,7 +56,7 @@ public class ComputerDTO {
 		return companyId;
 	}
 
-	public void setCompanyId(int companyId) {
+	public void setCompanyId(long companyId) {
 		this.companyId = companyId;
 	}
